rename `*-settings-panel` to `*-panel`
